Create Use the flex-direction Property to Make a Column
Use the flex-direction Property to Make a Column
The last two challenges used the flex-direction property set to row. This property can also create a column by vertically stacking the children of a flex container.

Add the CSS property flex-direction to the #box-container element, and give it a value of column.