Create Use the align-items Property in the Tweet Embed
Use the align-items Property in the Tweet Embed
The last challenge introduced the align-items property and gave an example. This property can be applied to a few tweet embed elements to align the flex items inside them.

Add the CSS property align-items to the header's .follow-btn element. Set the value to center.